Add single file module for proxy config
