+++ proxy_config/image_viewer_proxy.py
@@ -1,0 +1,17 @@
+from image_viewer import app_file
+
+
+def setup_proxy_config():
+    return {
+        "command": [
+            "panel",
+            "serve",
+            str(app_file),
+            "--allow-websocket-origin=*",
+            "--port",
+            "{port}",
+        ],
+        "launcher_entry": {
+            "enabled": False,
+        }
+    }
